# config/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Watch & Do API",
        "description": "Watch & Do API",
        "contact": {
            "name": "Alexandre Maia",
            "url": "http://wwwalexandremaia.pt",
            "email": "info@alexandremaia.pt"
        },
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/videos/{id}": {
            "get": {
                "tags": [ "Videos" ],
                "description": "<h2>Get Video Info by Id</h2><br><p>Return the selected Video with all associated Topics.</p></p>Return empty if no active Topic associated</p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video Info",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Videos" },
                                "examples": {
                                    "video": {
                                        "id": 4,
                                        "title": "teste title 4",
                                        "youtubeid": "testeyoutubeid 4",
                                        "Topics": [
                                            {
                                                "id": 3,
                                                "topic": "CSS",
                                                "description": "Altere a aparência de um Site",
                                                "TopicVideo": {
                                                    "createdAt": "2020-12-01T16:53:36.000Z",
                                                    "updatedAt": "2020-12-01T16:53:36.000Z",
                                                    "TopicId": 3,
                                                    "VideoId": 4
                                                }
                                            }
                                        ]
                                    }
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/videos/{id}/relations": {
            "get": {
                "tags": [ "Videos" ],
                "description": "<h2>Get Video Info by Id with relations</h2><br><p>Return the selected Video with all associated data.</p></p>Relations:</p><p>- Topics;</p><p>- Comments.</p></p>Return empty if no active Topic associated</p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video Info",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Videos" },
                                "examples": {
                                    "video": {
                                        "id": 4,
                                        "title": "teste title 4",
                                        "youtubeid": "testeyoutubeid 4",
                                        "Topics": [
                                            {
                                                "id": 3,
                                                "topic": "CSS",
                                                "description": "Altere a aparência de um Site",
                                                "TopicVideo": {
                                                    "createdAt": "2020-12-01T16:53:36.000Z",
                                                    "updatedAt": "2020-12-01T16:53:36.000Z",
                                                    "TopicId": 3,
                                                    "VideoId": 4
                                                }
                                            }
                                        ]
                                    },
                                    "comments": [
                                        {
                                            "id": 1,
                                            "userId": 1,
                                            "videoId": 4,
                                            "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum. Vestibulum dignissim sem et ligula mollis scelerisque. Donec faucibus et est sit amet cursus.",
                                            "status": 1,
                                            "archived": 0,
                                            "createdAt": "2020-12-05T10:57:37.000Z",
                                            "updatedAt": "2020-12-05T17:02:09.000Z",
                                            "UserId": 1,
                                            "VideoId": 4
                                        }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/wd/videos/{id}": {
            "get": {
                "tags": [ "Videos" ],
                "description": "<h2>Get Video Info by Id</h2><br><p>Role required: <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video Info",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Videos" },
                                "examples": {
                                    "Video": {
                                        "id": "1",
                                        "title": "HTML Crash Course For Absolute Beginners",
                                        "youtubeid": "UB1O30fR-EE",
                                        "status": "1"
                                    }
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "put": {
                "tags": [ "Videos" ],
                "description": "<h2>Update Video</h2><br><p>Role required: <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Videos" },
                            "examples": {
                                "Topic": {
                                    "value": {
                                        "title": "Updated title",
                                        "youtubeid": "UB1O30fR-EE",
                                        "topics" : [{"id": 32}, {"id": 6}]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Video",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Videos" }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "delete": {
                "tags": [ "Videos" ],
                "description": "<h2>Delete Video</h2><br><p>Role required: <strong>Super Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "Video Deleted"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/videos/{id}/relations": {
            "get": {
                "tags": [ "Videos" ],
                "description": "<h2>Get Video by Id with relations</h2><br><p>Relations:<p>-> Topics;</p><p>-> Comments;</p><p>-> Favorites with User relations.</p><br><p>Role required: <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Videos List",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Videos" },
                                "examples": {
                                    "Video List": [
                                        {
                                            "id": 1,
                                            "title": "teste title",
                                            "youtubeid": "testeyoutubeid",
                                            "status": true,
                                            "createdAt": "2020-12-01T16:13:30.000Z",
                                            "updatedAt": "2020-12-01T16:13:30.000Z",
                                            "Topics": [
                                                {
                                                    "id": 7,
                                                    "topic": "Updated Topic",
                                                    "description": "Updated Topic description",
                                                    "hexcolor": "#F1C40F",
                                                    "status": true,
                                                    "createdAt": "2020-11-29T22:24:41.000Z",
                                                    "updatedAt": "2020-11-29T22:24:41.000Z",
                                                    "TopicVideo": {
                                                        "createdAt": "2020-12-01T19:41:56.000Z",
                                                        "updatedAt": "2020-12-01T19:41:56.000Z",
                                                        "TopicId": 7,
                                                        "VideoId": 1
                                                    }
                                                }
                                            ],
                                            "Comments": [
                                                {
                                                    "id": 5,
                                                    "userId": 3,
                                                    "videoId": 1,
                                                    "comment": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin ut efficitur nisl. Phasellus porta porttitor pulvinar. Nunc scelerisque ullamcorper nibh a consectetur. Sed ac semper ipsum.",
                                                    "status": 1,
                                                    "archived": 0,
                                                    "createdAt": "2020-12-05T21:41:52.000Z",
                                                    "updatedAt": "2020-12-05T21:41:52.000Z"
                                                }
                                            ],
                                            "Favorites": [
                                              {
                                                    "id": 7,
                                                    "userId": 1,
                                                    "videoId": 1,
                                                    "createdAt": "2020-12-04T17:15:19.000Z",
                                                    "updatedAt": "2020-12-04T17:15:19.000Z",
                                                    "UserId": 1,
                                                    "VideoId": 1,
                                                    "User": {
                                                        "id": 1,
                                                        "username": "supadmin",
                                                        "email": "supadmin@mail.pt",
                                                        "password": "$2b$10$CKL60nmJrmaFGTcxjEN5F.xQKs/YHkE4wwd5c8aFrJvwE4l5/rlpS",
                                                        "status": true,
                                                        "role_id": 3,
                                                        "createdAt": "2020-11-29T09:23:05.000Z",
                                                        "updatedAt": "2020-11-29T09:23:05.000Z"
                                                    }
                                                }
                                            ]
                                        }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
...
    }
}